// src/pages/AdminProperties/models/properties.model.ts
export interface PropertiesForm {
  document: string;
  name: string;
  phone: string;
  address: string;
  email: string;
  percentageCommission: string;
  accountNumber: string;
  accountType: string;
  bank: string;
  fourPerThousand: boolean;
  fourPerThousandValue: number;
  property: string[];
  administrationPayment: string;
}